api/requirements.txt
azure-functions
Flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
sendgrid==6.11.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
werkzeug==3.1.3
